backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
boto3==1.34.0
openpyxl==3.1.2
pypdf==5.1.0
python-docx==1.1.2
reportlab==4.1.0
anthropic==0.26.0
openai==1.55.0
pytest==8.2.0
pytest-asyncio==0.23.6
